Handle an empty tabstrip in TabStrip::GetDropBounds.
(cherry picked from commit 8c956ce0d96372940ec1702bbacec4b37cfa5357)

Bug: 1209616
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3017667
Commit-Queue: Taylor Bergquist <tbergquist@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#900744}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3061458
Commit-Queue: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/branch-heads/4515@{#1918}
Cr-Branched-From: 488fc70865ddaa05324ac00a54a6eb783b4bc41c-refs/heads/master@{#885287}

+++ chrome/browser/ui/views/tabs/tab_strip.cc
@@ -3398,6 +3398,12 @@ gfx::Rect TabStrip::GetDropBounds(int drop_index,
   // The X location the indicator points to.
   int center_x = -1;
 
+  if (GetTabCount() == 0) {
+    // If the tabstrip is empty, it doesn't matter where the drop arrow goes.
+    // The tabstrip can only be transiently empty, e.g. during shutdown.
+    return gfx::Rect();
+  }
+
   Tab* tab = tab_at(std::min(drop_index, GetTabCount() - 1));
   const bool first_in_group =
       drop_index < GetTabCount() && tab->group().has_value() &&
